Read deps from Puppetfiles in git repos
It is rather annoying to have to release a module just so you can have it depend
on other modules.  We've got an internal tree of modules that we can easily
reference between, and it would be nice to be able to use Puppetfiles inside those
module trees to do the dereferencing.  Well, now you can.