Adding the ability to use rsync to converge cached and installed modules
Need to use the parent of dest for rsync

Reverting gemfile

Adding a test. Going to check with maintainers to see if this is a valid testing strategy.

Adding ctime to the mix to make sure that the file isn't just assigned the same inode

Adding tests for the rsync option using the install command

Only compare ctime if inodes match

Adding test case to make sure update works properly with git